OPHJOD-1219: Refactor duplicate MahdollisuusTyyppi
